+++ src/main/java/roomescape/payment/processor/PaymentProcessor.java
@@ -1,5 +1,8 @@
 package roomescape.payment.processor;
 
+import roomescape.payment.dto.PaymentConfirmRequest;
+import roomescape.payment.dto.PaymentConfirmResponse;
+
 public interface PaymentProcessor {
     boolean supports(final PaymentType paymentType);
