+++ src/main/java/roomescape/payment/processor/toss/TossPaymentConfirmRequest.java
@@ -1,6 +1,6 @@
 package roomescape.payment.processor.toss;
 
-import roomescape.payment.processor.PaymentConfirmRequest;
+import roomescape.payment.dto.PaymentConfirmRequest;
 
 public record TossPaymentConfirmRequest(int amount, String orderId, String paymentKey) implements
         PaymentConfirmRequest {
