+++ src/main/java/roomescape/payment/processor/toss/TossPaymentConfirmResponse.java
@@ -1,6 +1,6 @@
 package roomescape.payment.processor.toss;
 
-import roomescape.payment.processor.PaymentConfirmResponse;
+import roomescape.payment.dto.PaymentConfirmResponse;
 
 public record TossPaymentConfirmResponse(String orderId, String paymentKey)
         implements PaymentConfirmResponse {
