+++ src/main/java/roomescape/reservation/dto/CreateReservationRequest.java
@@ -1,7 +1,7 @@
 package roomescape.reservation.dto;
 
 import java.time.LocalDate;
-import roomescape.payment.processor.PaymentConfirmRequest;
+import roomescape.payment.dto.PaymentConfirmRequest;
 import roomescape.payment.processor.PaymentType;
 
 public record CreateReservationRequest(
