+++ src/test/java/roomescape/admin/controller/AdminRestControllerTest.java
@@ -12,7 +12,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.context.jdbc.Sql;
-import roomescape.payment.processor.PaymentConfirmRequest;
+import roomescape.payment.dto.PaymentConfirmRequest;
 import roomescape.payment.processor.PaymentType;
 import roomescape.payment.processor.toss.TossPaymentConfirmRequest;
 import roomescape.payment.processor.toss.TossPaymentProcessor;
